+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - 105 Job applications
+- 109 Job alert subscription
+- 110 Job search alert notifications
+- 111 Job alert unsubscribe
 
 ## 4.0 (2020-02-24)